// types/events.types.ts
export enum EventTimeFilter {
  Any = 'Any',
  Upcoming = 'Upcoming',
  Past = 'Past',
}

export enum EventLocationFilter {
  All = 'All',
  Online = 'Online',
  Offline = 'Offline',
}

export enum EventStatusFilter {
  Any = 'Any',
  Open = 'Open',
  Closed = 'Closed',
}

export enum EventTypeFilter {
  All = 'All',
  Economic = 'Economic',
  Social = 'Social',
  Environmental = 'Environmental',
}